requirements.txt
Django==3.1
django-model-utils==4.0.0
django-bootstrap4==2.3.1
django-crispy-forms==1.10.0